Move common `Application` interface into separate file
